Moved "util" package contents into correct "utils" package
