+++ src/com/redstoner/modules/blockplacemods/BlockPlaceMods.java
@@ -1,5 +1,16 @@
 package com.redstoner.modules.blockplacemods;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Comparator;
+import java.util.List;
+import java.util.Map;
+
+import org.bukkit.ChatColor;
+import org.bukkit.command.CommandSender;
+import org.bukkit.entity.Player;
+import org.bukkit.event.Listener;
+
 import com.nemez.cmdmgr.Command;
 import com.redstoner.annotations.AutoRegisterListener;
 import com.redstoner.annotations.Version;
@@ -9,17 +20,11 @@ import com.redstoner.modules.Module;
 import com.redstoner.modules.blockplacemods.mods.Mod;
 import com.redstoner.modules.blockplacemods.mods.ModAbstract;
 import com.redstoner.modules.blockplacemods.mods.ModToggledAbstract;
-import com.redstoner.modules.blockplacemods.util.CommandException;
-import com.redstoner.modules.blockplacemods.util.CommandMap;
-import org.bukkit.ChatColor;
-import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
-import org.bukkit.event.Listener;
-
-import java.util.*;
+import com.redstoner.utils.CommandException;
+import com.redstoner.utils.CommandMap;
 
 @AutoRegisterListener
-@Version(major = 3, minor = 2, revision = 3, compatible = 3)
+@Version(major = 3, minor = 2, revision = 4, compatible = 3)
 public final class BlockPlaceMods implements Module, Listener
 {
 	public static String PREFIX = ChatColor.GRAY + "[" + ChatColor.DARK_GREEN + "BPM" + ChatColor.GRAY + "]"
@@ -90,7 +95,6 @@ public final class BlockPlaceMods implements Module, Listener
 	 * "}\n" +
 	 * "}\n";
 	 * } */
-	
 	@Command(hook = "mod_empty")
 	public void onModEmptyCommand(CommandSender sender)
 	{
@@ -104,34 +108,46 @@ public final class BlockPlaceMods implements Module, Listener
 				.filter(x -> x != null && !x.trim().isEmpty()).toArray(String[]::new);
 		String prefix = PREFIX;
 		String message;
-		
-		try {
-			
-			if (args.length > 0) {
+		try
+		{
+			if (args.length > 0)
+			{
 				Mod target = ModAbstract.getMod(args[0].toLowerCase());
-				if (target != null) {
+				if (target != null)
+				{
 					prefix += "&7[&2" + capitalize(target.getName()) + "&7]&a";
-					if (!(sender instanceof Player)) {
+					if (!(sender instanceof Player))
+					{
 						message = "&cYou must be a player to use any block place mod";
-					} else {
+					}
+					else
+					{
 						message = target.runCommand((Player) sender, Arrays.copyOfRange(args, 1, args.length));
 					}
-				} else if (args[0].equalsIgnoreCase("help")) {
+				}
+				else if (args[0].equalsIgnoreCase("help"))
+				{
 					message = commandHelp(sender, args);
-				} else {
+				}
+				else
+				{
 					message = "&cThat argument could not be recognized";
 				}
-			} else {
+			}
+			else
+			{
 				message = commandHelp(sender, args);
 			}
-			
-		} catch (CommandException ex) {
+		}
+		catch (CommandException ex)
+		{
 			message = " &c" + ex.getMessage();
-		} catch (Throwable t) {
+		}
+		catch (Throwable t)
+		{
 			message = " &cAn unexpected error occurred while executing this command.";
 			t.printStackTrace();
 		}
-		
 		Utils.sendMessage(sender, prefix, message, '&');
 	}
 	
@@ -139,36 +155,35 @@ public final class BlockPlaceMods implements Module, Listener
 	{
 		StringBuilder result = new StringBuilder("§7BlockPlaceMods adds some redstone-centric utilities");
 		result.append("\n").append(ChatColor.GRAY.toString()).append("Available mods:");
-		
 		List<Mod> mods = new ArrayList<>(ModAbstract.getMods().values());
-		mods.sort(Comparator.<Mod>comparingInt(m -> ModToggledAbstract.class.isInstance(m) ? 1 : -1).thenComparing(Mod::getName));
-		
+		mods.sort(Comparator.<Mod> comparingInt(m -> ModToggledAbstract.class.isInstance(m) ? 1 : -1)
+				.thenComparing(Mod::getName));
 		for (Mod mod : mods)
 		{
 			result.append("\n").append(ChatColor.AQUA.toString()).append("/mod ").append(ChatColor.ITALIC.toString())
 					.append(mod.getName());
-			
-			for (String alias : mod.getAliases()) {
+			for (String alias : mod.getAliases())
+			{
 				result.append('|').append(alias);
 			}
-			
-			result.append(ChatColor.GRAY.toString()).append(" - ")
-					.append(mod.getDescription());
+			result.append(ChatColor.GRAY.toString()).append(" - ").append(mod.getDescription());
 		}
 		return result.toString();
 	}
 	
-	private static String capitalize(String modName) {
-		if (modName.isEmpty()) {
+	private static String capitalize(String modName)
+	{
+		if (modName.isEmpty())
+		{
 			return modName;
 		}
 		char first = modName.charAt(0);
-		if (first != (first = Character.toUpperCase(first))) {
+		if (first != (first = Character.toUpperCase(first)))
+		{
 			char[] result = modName.toCharArray();
 			result[0] = first;
 			return String.valueOf(result);
 		}
 		return modName;
 	}
-	
 }
